+++ protoc_plugin/lib/src/gen/google/protobuf/dart_edition_defaults.pb.dart
@@ -1,0 +1,9 @@
+// Generated with third_party/dart/protoc_plugin/tool/regenerate.sh.
+// Copyright (c) 2013, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// ignore_for_file: constant_identifier_names
+
+const ProtobufInternalDartEditionDefaults =
+    'ChcYhAciACoQCAEQAhgCIAMoATACOAJAAQoXGOcHIgAqEAgCEAEYASACKAEwATgCQAEKFxjoByIMCAEQARgBIAIoATABKgQ4AkABChcY6QciEAgBEAEYASACKAEwATgBQAIqACDmByjpBw==';