+++ protoc_plugin/CHANGELOG.md
@@ -6,9 +6,11 @@ Note: this version requires protobuf 5.0.0.
 * Update generated `clone` members to take advantage of faster `deepCopy`
   implementation in protobuf 5.0.0. ([#742])
 * Code size improvements for enum fields. ([#1047])
+* Support protobuf editions. ([#1052])
 
 [#742]: https://github.com/google/protobuf.dart/pull/742
 [#1047]: https://github.com/google/protobuf.dart/pull/1047
+[#1052]: https://github.com/google/protobuf.dart/pull/1052
 
 ## 22.5.0